Modify struct names and gateways to be compatible with other Adyen libs
