+++ payment.go
@@ -20,17 +20,3 @@ type AuthoriseResponse struct {
 type AdditionalData struct {
 	Content string `json:"card.encrypted.json"`
 }
-
-// Capture structure for Capture request
-type Capture struct {
-	ModificationAmount *Amount `json:"modificationAmount"`
-	Reference          string  `json:"reference"`
-	MerchantAccount    string  `json:"merchantAccount"`
-	OriginalReference  string  `json:"originalReference"`
-}
-
-// CaptureResponse is a response structure for Adyen capture
-type CaptureResponse struct {
-	PspReference string `json:"pspReference"`
-	Response     string `json:"response"`
-}
